+++ pulsar/apps/test/__init__.py
@@ -191,8 +191,8 @@ configuration and plugins.'''
             self.runner.on_start()
             monitor.cfg.set('workers',min(self.cfg.workers,len(tests)))
             self._time_start = time.time()
-            for _,testcls in tests:
-                monitor.put(TestRequest(testcls))
+            for tag,testcls in tests:
+                monitor.put(TestRequest(testcls,tag))
         else:
             print('Could not find any tests.')
             monitor.arbiter.stop()